seperate tests 'sanity' and 'minimal-config' vs all
